Fix: Flag automated transfer as active on attempt not confirmation
Previously we were waiting to hear back from the WordPress.com servers
for the "initiate" request to start an automated transfer before we
flagged a given site as having actually started the transfer.

Now we will watch the requst action instead of the success action for
making this decision, meaning that the state should change immediately
after starting the transfer in Calypso.

If the request fails, we will reset the state to indicate that we have
queried the eligibility endpoint but aren't transfering yet.